backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
loguru==0.7.2
orjson==3.10.7
pandas==2.2.2
openpyxl==3.1.5
python-multipart==0.0.19
